totvs/requirements.txt
playwright>=1.44
python-dotenv>=1.0
openpyxl>=3.1
pandas>=2.2
